move helper functions to BaseMutation
