+++ apps/central-api/src/graphql/mutation/OrderMutation.ts
@@ -6,25 +6,23 @@ import * as _ from 'lodash'
 import { EntityManager, In } from 'typeorm'
 
 import { Menu, Order, OrderItem, OrderItemStatusEnum, OrderType } from '@db/entities'
-import { ValidationHelper } from '@shared/helpers'
 import { z } from 'zod/v4'
 import { BaseMutArgs, BaseMutation } from './BaseMutation'
 
-const zItemSchema = z.object({
-  id: z.string().optional(),
+const valuesSchema = z.object({
   deletedAt: z.boolean().optional(),
-  menuId: z.string(),
-  quantity: z.number(),
-  orderId: z.string().optional(),
+  orderItems: z.array(
+    z.object({
+      id: z.string().optional(),
+      deletedAt: z.boolean().optional(),
+      menuId: z.string(),
+      quantity: z.number(),
+      orderId: z.string().optional(),
+    }),
+  ),
 })
 
-const zValues = z.object({
-  deletedAt: z.boolean().optional(),
-  orderItems: z.array(zItemSchema),
-})
-
-type TItem = z.infer<typeof zItemSchema>
-type TValues = z.infer<typeof zValues>
+type TValues = z.infer<typeof valuesSchema>
 
 @Resolver(() => OrderType)
 export class OrderMutationResolver extends BaseMutation {
@@ -35,16 +33,14 @@ export class OrderMutationResolver extends BaseMutation {
 
   @Mutation(() => OrderType, { name: 'OrderMutation' })
   async resolve(@Args() args: BaseMutArgs, @Context() context: MyGraphQlContext) {
-    ValidationHelper.validate(args.values, zValues)
+    this.validate(args.values, valuesSchema)
 
     const { deletedAt, ...values } = args.values as TValues
     const { id } = args
 
     return this.dbService.withTransaction(context.merchantId, async (db) => {
       this.db = db
-      const order = id
-        ? await this.db.findOneOrFail(Order, { where: { id }, relations: { orderItems: true } })
-        : this.db.create(Order)
+      const order = await this.findOneOrCreate<Order>(db, Order, { where: { id } })
 
       if (id && deletedAt) return await order.softRemove()
 
@@ -52,7 +48,7 @@ export class OrderMutationResolver extends BaseMutation {
       order.staffId = context.staffId
 
       await order.save()
-      const updatedOrder = await this.db.findOneOrFail(Order, { where: { id }, relations: { orderItems: true } })
+      const updatedOrder = await this.db.findOneOrFail(Order, { where: { id: order.id }, relations: { orderItems: true } })
 
       if (values.orderItems.length) {
         await this.saveOrderItems(updatedOrder, values.orderItems)
@@ -62,15 +58,13 @@ export class OrderMutationResolver extends BaseMutation {
     })
   }
 
-  // if a menu is already in order, merge the quantity
-  async saveOrderItems(order: Order, items: TItem[] = []) {
+  async saveOrderItems(order: Order, items: TValues['orderItems'] = []) {
     const menus = await this.db.findBy(Menu, { id: In(_.map(items, 'menuId')) })
     const dbOrderItems = order.orderItems
 
     for (const item of items) {
       const dbOrderItem = _.find(dbOrderItems, { menuId: item.menuId })
-      const orderItem =
-        dbOrderItem || (item.id ? await this.db.findOneByOrFail(OrderItem, { id: item.id }) : this.db.create(OrderItem))
+      const orderItem = dbOrderItem || (await this.findOneOrCreate<OrderItem>(this.db, OrderItem, { where: { id: item.id } }))
       const menu = _.find(menus, { id: item.menuId })
 
       if (!menu) throw new HttpException('Menu not found', HttpStatus.BAD_REQUEST)
@@ -80,6 +74,7 @@ export class OrderMutationResolver extends BaseMutation {
         continue
       }
 
+      // if a menu is already in order, merge the quantity
       if (dbOrderItem) {
         orderItem.fill({ quantity: dbOrderItem.quantity + item.quantity })
       } else {
