+++ apps/central-api/src/shared/exceptions/index.ts
@@ -1,1 +1,1 @@
-export * from './validation.exception'
+export * from './GqlValidationEx'
