--- apps/central-api/src/shared/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './ValidationHelper'
